use provided cityName to search for weather data
this part is a little tricky to understand for me
right now. the onPressed in which this code lives
is basically getting a return value from the screen
it loads. the weatherData search part is on the
tail end of this loop, when the next screen has
returned the city name and this location screen is
now ready to lookup the api for that city.

i can't explain it better right now. later