require the INTERNET permission
how else are we going to make the api call ;)

missing this was the reason the app was failing in
release builds.

no idea how debug builds handle internet permissions
but it seems it now works without having added this
crucial bit unlike in older versions when it required
this permission merely for the debug bridge with the
computer to work.

hah! this has caused a lot of hair-pulling, and
hair loss.